requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
mysql-connector==2.2.9